feat: 팀 참여하기 기능 완성 (FE-80)
