+++ src/constants/errorMessage.ts
@@ -1,6 +1,9 @@
-const ErrorMessages = {
+const ERROR_MESSAGES = {
   PROFILE_REQUIRED: "프로필 이미지를 넣어주세요.",
   INVALID_TEAMNAME: "이미 존재하는 이름입니다.",
+  INVALID_FORMAT: "올바르지 않은 형식입니다.",
+  NON_EXISTENT_GROUP: "존재하지 않는 그룹 초대 링크입니다.",
+  ALREADY_MEMBER: "이미 속해 있는 그룹입니다.",
 };
 
-export default ErrorMessages;
+export default ERROR_MESSAGES;
