+++ src/utils/getParams.ts
@@ -1,0 +1,6 @@
+export const getParams = (link: string) => {
+  const url = new URL(link);
+  const params = new URLSearchParams(url.search);
+
+  return params;
+};
