+++ src/utils/validateImage.ts
@@ -1,0 +1,2 @@
+export const validateImage = (imageUrl: string | null) =>
+  imageUrl?.includes("sprint") ? imageUrl : null;
